Add test for complex options
Instead of only allowing boolean and with some exceptions numeric
values as option value, complex options are parsed like a simple
version of JSON. They are needed in order to define values in groups
and with this to get things like #28 work properly.
Here are the tests for the complex options. They are only allowed in
/*jshint - comments to not provoke more errors in jslint as needed.