Remove tab from unsafe characters
Before jshint tests for unsafe characters (in "nextline()"), every tab
was replaced with blanks (option indent pretends, how many blanks).

So removing tab from the test for unsafe characters is surely safe.

Instead of removing it there, the tab is now recognized in the test
for correct indentation.
This change is needed to allow tabs in the new format-test
implementation for those who want to use tab as well as don't allow
constructs like
   function\t(arg)\t{
   }
Currently this is allowed - even with option white=true.